+++ MapboxVision/Services/ModelPerformance.swift
@@ -4,7 +4,15 @@ import UIKit
 /**
  Enumeration which determines whether SDK should adapt its performance to environmental changes (acceleration/deceleration, standing time) or stay fixed.
  */
-public enum ModelPerformanceMode {
+public enum ModelPerformanceMode: Comparable {
+    /// :nodoc:
+    public static func < (lhs: ModelPerformanceMode, rhs: ModelPerformanceMode) -> Bool {
+        if lhs == rhs { return false }
+        // We condider `dynamic` to be less important than `fixed` when encountering both.
+        // That's because `dynamic` mode may cause a `low` performance rate, while `fixed` one tries to keed a set value.
+        return lhs == .dynamic
+    }
+
     /**
      Fixed mode.
      */
@@ -19,7 +27,16 @@ public enum ModelPerformanceMode {
 /**
  Enumeration which determines performance rate of the specific model. These are high-level settings that translates into adjustment of FPS for ML model inference.
  */
-public enum ModelPerformanceRate {
+public enum ModelPerformanceRate: Comparable, CaseIterable {
+    /// :nodoc:
+    public static func < (lhs: ModelPerformanceRate, rhs: ModelPerformanceRate) -> Bool {
+        let rates = ModelPerformanceRate.allCases
+        guard let lhsIndex = rates.firstIndex(of: lhs), let rhsIndex = rates.firstIndex(of: rhs) else {
+            fatalError("Rates array is not complete")
+        }
+        return lhsIndex < rhsIndex
+    }
+
     /**
      Identifies that output of particular model is not required.
      */
@@ -44,6 +61,7 @@ public enum ModelPerformanceRate {
 /**
  Enumeration representing configuration for ML models
  */
+@available(*, deprecated, message: "Use ModelPerformance structure instead. Enum will be removed in version 0.13.0")
 public enum ModelPerformanceConfig: Equatable {
     /// :nodoc:
     public static func == (lhs: ModelPerformanceConfig, rhs: ModelPerformanceConfig) -> Bool {
@@ -76,7 +94,15 @@ public enum ModelPerformanceConfig: Equatable {
 /**
  Structure representing performance setting for tasks related to specific ML model. It’s defined as a combination of mode and rate.
  */
-public struct ModelPerformance: Equatable {
+public struct ModelPerformance: Comparable {
+    /// :nodoc:
+    public static func < (lhs: ModelPerformance, rhs: ModelPerformance) -> Bool {
+        if lhs.mode == rhs.mode {
+            return lhs.rate < rhs.rate
+        }
+        return lhs.mode < rhs.mode
+    }
+
     /**
      Performance Mode.
      */
@@ -123,27 +149,11 @@ enum ModelPerformanceResolver {
 
     private static let isHighPerformance = UIDevice.current.isHighPerformance
 
-    private static let    segmentationHighEnd = PerformanceEntry(off: 1, low: 2, high: 7)
-    private static let       detectionHighEnd = PerformanceEntry(off: 3, low: 4, high: 12)
     private static let mergedSegDetectHighEnd = PerformanceEntry(off: 3, low: 4, high: 12)
-
-    private static let     segmentationLowEnd = PerformanceEntry(off: 1, low: 2, high: 5)
-    private static let        detectionLowEnd = PerformanceEntry(off: 3, low: 4, high: 11)
     private static let  mergedSegDetectLowEnd = PerformanceEntry(off: 3, low: 4, high: 11)
 
-    private static func performanceEntry(for model: MLModelType) -> PerformanceEntry {
-        switch model {
-        case .segmentation:
-            return isHighPerformance ? segmentationHighEnd : segmentationLowEnd
-        case .detection:
-            return isHighPerformance ? detectionHighEnd : detectionLowEnd
-        case .mergedSegDetect:
-            return isHighPerformance ? mergedSegDetectHighEnd : mergedSegDetectLowEnd
-        }
-    }
-
-    static func coreModelPerformance(for model: MLModelType, with performance: ModelPerformance) -> CoreModelPerformance {
-        let entry = performanceEntry(for: model)
+    static func coreModelPerformance(with performance: ModelPerformance) -> CoreModelPerformance {
+        let entry = isHighPerformance ? mergedSegDetectHighEnd : mergedSegDetectLowEnd
 
         switch performance.mode {
         case .fixed: